API

equation_search

SymbolicRegression.equation_search — Method

equation_search(X, y[; kws...])

Perform a distributed equation search for functions f_i which describe the mapping f_i(X[:, j]) ≈ y[i, j]. Options are configured using SymbolicRegression.Options(...), which should be passed as a keyword argument to options. One can turn off parallelism with numprocs=0, which is useful for debugging and profiling.

Arguments

X::AbstractMatrix{T}: The input dataset to predict y from. The first dimension is features, the second dimension is rows.

y::Union{AbstractMatrix{T}, AbstractVector{T}}: The values to predict. The first dimension is the output feature to predict with each equation, and the second dimension is rows.

niterations::Int=10: The number of iterations to perform the search. More iterations will improve the results.

weights::Union{AbstractMatrix{T}, AbstractVector{T}, Nothing}=nothing: Optionally weight the loss for each y by this value (same shape as y).

variable_names::Union{Vector{String}, Nothing}=nothing: The names of each feature in X, which will be used during printing of equations.

options::Options=Options(): The options for the search, such as which operators to use, evolution hyperparameters, etc.

parallelism=:multithreading: What parallelism mode to use. The options are :multithreading, :multiprocessing, and :serial. By default, multithreading will be used. Multithreading uses less memory, but multiprocessing can handle multi-node compute. If using :multithreading mode, the number of threads available to julia are used. If using :multiprocessing, numprocs processes will be created dynamically if procs is unset. If you have already allocated processes, pass them to the procs argument and they will be used. You may also pass a string instead of a symbol, like "multithreading".

numprocs::Union{Int, Nothing}=nothing: The number of processes to use, if you want equation_search to set this up automatically. By default this will be 4, but can be any number (you should pick a number <= the number of cores available).

procs::Union{Vector{Int}, Nothing}=nothing: If you have set up a distributed run manually with procs = addprocs() and @everywhere, pass the procs to this keyword argument.

addprocs_function::Union{Function, Nothing}=nothing: If using multiprocessing (parallelism=:multithreading), and are not passing procs manually, then they will be allocated dynamically using addprocs. However, you may also pass a custom function to use instead of addprocs. This function should take a single positional argument, which is the number of processes to use, as well as the lazy keyword argument. For example, if set up on a slurm cluster, you could pass addprocs_function = addprocs_slurm, which will set up slurm processes.

runtests::Bool=true: Whether to run (quick) tests before starting the search, to see if there will be any problems during the equation search related to the host environment.

saved_state::Union{StateType, Nothing}=nothing: If you have already run equation_search and want to resume it, pass the state here. To get this to work, you need to have set returnstate=true, which will cause `equationsearch` to return the state. Note that you cannot change the operators or dataset, but most other options should be changeable.

return_state::Union{Bool, Nothing}=nothing: Whether to return the state of the search for warm starts. By default this is false.

loss_type::Type=Nothing: If you would like to use a different type for the loss than for the data you passed, specify the type here. Note that if you pass complex data ::Complex{L}, then the loss type will automatically be set to L.

Returns

hallOfFame::HallOfFame: The best equations seen during the search. hallOfFame.members gives an array of PopMember objects, which have their tree (equation) stored in .tree. Their score (loss) is given in .score. The array of PopMember objects is enumerated by size from 1 to options.maxsize.

Options

Missing docstring.

Missing docstring for Options(;). Check Documenter's build log for details.

SymbolicRegression.CoreModule.OptionsStructModule.MutationWeights — Method

MutationWeights(;kws...)

This defines how often different mutations occur. These weightings will be normalized to sum to 1.0 after initialization.

Arguments

mutate_constant::Float64: How often to mutate a constant.

mutate_operator::Float64: How often to mutate an operator.

add_node::Float64: How often to append a node to the tree.

insert_node::Float64: How often to insert a node into the tree.

delete_node::Float64: How often to delete a node from the tree.

simplify::Float64: How often to simplify the tree.

randomize::Float64: How often to create a random tree.

do_nothing::Float64: How often to do nothing.

optimize::Float64: How often to optimize the constants in the tree, as a mutation. Note that this is different from optimizer_probability, which is performed at the end of an iteration for all individuals.

Printing

DynamicExpressions.EquationModule.string_tree — Method

string_tree(tree::Node, options::Options; kws...)

Convert an equation to a string.

Arguments

tree::Node: The equation to convert to a string.

options::Options: The options holding the definition of operators.

variable_names::Union{Array{String, 1}, Nothing}=nothing: what variables to print for each feature.

Evaluation

DynamicExpressions.EvaluateEquationModule.eval_tree_array — Method

eval_tree_array(tree::Node, X::AbstractArray, options::Options; kws...)

Evaluate a binary tree (equation) over a given input data matrix. The operators contain all of the operators used. This function fuses doublets and triplets of operations for lower memory usage.

This function can be represented by the following pseudocode:

function eval(current_node)
    if current_node is leaf
        return current_node.value
    elif current_node is degree 1
        return current_node.operator(eval(current_node.left_child))
    else
        return current_node.operator(eval(current_node.left_child), eval(current_node.right_child))

The bulk of the code is for optimizations and pre-emptive NaN/Inf checks, which speed up evaluation significantly.

Arguments

tree::Node: The root node of the tree to evaluate.

X::AbstractArray: The input data to evaluate the tree on.

options::Options: Options used to define the operators used in the tree.

Returns

(output, complete)::Tuple{AbstractVector, Bool}: the result, which is a 1D array, as well as if the evaluation completed successfully (true/false). A false complete means an infinity or nan was encountered, and a large loss should be assigned to the equation.

Derivatives

SymbolicRegression.jl can automatically and efficiently compute derivatives of expressions with respect to variables or constants. This is done using either eval_diff_tree_array, to compute derivative with respect to a single variable, or with eval_grad_tree_array, to compute the gradient with respect all variables (or, all constants). Both use forward-mode automatic, but use Zygote.jl to compute derivatives of each operator, so this is very efficient.

DynamicExpressions.EvaluateEquationDerivativeModule.eval_diff_tree_array — Method

eval_diff_tree_array(tree::Node, X::AbstractArray, options::Options, direction::Int)

Compute the forward derivative of an expression, using a similar structure and optimization to evaltreearray. direction is the index of a particular variable in the expression. e.g., direction=1 would indicate derivative with respect to x1.

Arguments

tree::Node: The expression tree to evaluate.

X::AbstractArray: The data matrix, with each column being a data point.

options::Options: The options containing the operators used to create the tree. enable_autodiff must be set to true when creating the options. This is needed to create the derivative operations.

direction::Int: The index of the variable to take the derivative with respect to.

Returns

(evaluation, derivative, complete)::Tuple{AbstractVector, AbstractVector, Bool}: the normal evaluation, the derivative, and whether the evaluation completed as normal (or encountered a nan or inf).

DynamicExpressions.EvaluateEquationDerivativeModule.eval_grad_tree_array — Method

eval_grad_tree_array(tree::Node, X::AbstractArray, options::Options; variable::Bool=false)

Compute the forward-mode derivative of an expression, using a similar structure and optimization to evaltreearray. variable specifies whether we should take derivatives with respect to features (i.e., X), or with respect to every constant in the expression.

Arguments

tree::Node: The expression tree to evaluate.

X::AbstractArray: The data matrix, with each column being a data point.

options::Options: The options containing the operators used to create the tree. enable_autodiff must be set to true when creating the options. This is needed to create the derivative operations.

variable::Bool: Whether to take derivatives with respect to features (i.e., X - with variable=true), or with respect to every constant in the expression (variable=false).

Returns

(evaluation, gradient, complete)::Tuple{AbstractVector, AbstractArray, Bool}: the normal evaluation, the gradient, and whether the evaluation completed as normal (or encountered a nan or inf).

SymbolicUtils.jl interface

DynamicExpressions.ExtensionInterfaceModule.node_to_symbolic — Method

node_to_symbolic(tree::Node, options::Options; kws...)

Convert an expression to SymbolicUtils.jl form.

Note that use of this function requires SymbolicUtils.jl to be installed and loaded.

Pareto frontier

SymbolicRegression.HallOfFameModule.calculate_pareto_frontier — Method

calculate_pareto_frontier(hallOfFame::HallOfFame{T,L}) where {T<:DATA_TYPE,L<:LOSS_TYPE}