API

EquationSearch

SymbolicRegression.EquationSearch — Method

EquationSearch(X, y[; kws...])

Perform a distributed equation search for functions f_i which describe the mapping f_i(X[:, j]) ≈ y[i, j]. Options are configured using SymbolicRegression.Options(...), which should be passed as a keyword argument to options. One can turn off parallelism with numprocs=0, which is useful for debugging and profiling.

Arguments

X::AbstractMatrix{T}: The input dataset to predict y from. The first dimension is features, the second dimension is rows.

y::Union{AbstractMatrix{T}, AbstractVector{T}}: The values to predict. The first dimension is the output feature to predict with each equation, and the second dimension is rows.

niterations::Int=10: The number of iterations to perform the search. More iterations will improve the results.

weights::Union{AbstractMatrix{T}, AbstractVector{T}, Nothing}=nothing: Optionally weight the loss for each y by this value (same shape as y).

varMap::Union{Vector{String}, Nothing}=nothing: The names of each feature in X, which will be used during printing of equations.

options::Options=Options(): The options for the search, such as which operators to use, evolution hyperparameters, etc.

parallelism=:multithreading: What parallelism mode to use. The options are :multithreading, :multiprocessing, and :serial. By default, multithreading will be used. Multithreading uses less memory, but multiprocessing can handle multi-node compute. If using :multithreading mode, the number of threads available to julia are used. If using :multiprocessing, numprocs processes will be created dynamically if procs is unset. If you have already allocated processes, pass them to the procs argument and they will be used.

numprocs::Union{Int, Nothing}=nothing: The number of processes to use, if you want EquationSearch to set this up automatically. By default this will be 4, but can be any number (you should pick a number <= the number of cores available).

procs::Union{Vector{Int}, Nothing}=nothing: If you have set up a distributed run manually with procs = addprocs() and @everywhere, pass the procs to this keyword argument.

addprocs_function::Union{Function, Nothing}=nothing: If using multiprocessing (parallelism=:multithreading), and are not passing procs manually, then they will be allocated dynamically using addprocs. However, you may also pass a custom function to use instead of addprocs. This function should take a single positional argument, which is the number of processes to use, as well as the lazy keyword argument. For example, if set up on a slurm cluster, you could pass addprocs_function = addprocs_slurm, which will set up slurm processes.

runtests::Bool=true: Whether to run (quick) tests before starting the search, to see if there will be any problems during the equation search related to the host environment.

saved_state::Union{StateType, Nothing}=nothing: If you have already run EquationSearch and want to resume it, pass the state here. To get this to work, you need to have stateReturn=true in the options, which will cause EquationSearch to return the state. Note that you cannot change the operators or dataset, but most other options should be changeable.

Returns

hallOfFame::HallOfFame: The best equations seen during the search. hallOfFame.members gives an array of PopMember objects, which have their tree (equation) stored in .tree. Their score (loss) is given in .score. The array of PopMember objects is enumerated by size from 1 to options.maxsize.

Options

Missing docstring.

Missing docstring for Options(; binary_operators::NTuple{nbin, Any}=(div, plus, mult), unary_operators::NTuple{nuna, Any}=(exp, cos), bin_constraints=nothing, una_constraints=nothing, ns=10, #1 sampled from every ns per mutation topn=10, #samples to return per population parsimony=0.000100f0, alpha=0.100000f0, maxsize=20, maxdepth=nothing, fast_cycle=false, migration=true, hofMigration=true, fractionReplacedHof=0.1f0, shouldOptimizeConstants=true, hofFile=nothing, npopulations=nothing, nrestarts=3, perturbationFactor=1.000000f0, annealing=true, batching=false, batchSize=50, mutationWeights=[10.000000, 1.000000, 1.000000, 3.000000, 3.000000, 0.010000, 1.000000, 1.000000], warmupMaxsize=0, useFrequency=false, npop=1000, ncyclesperiteration=300, fractionReplaced=0.1f0, verbosity=convert(Int, 1e9), probNegate=0.01f0, seed=nothing ) where {nuna,nbin}. Check Documenter's build log for details.

Printing

Missing docstring.

Missing docstring for string_tree(tree::Node, options::Options). Check Documenter's build log for details.

Evaluation

Missing docstring.

Missing docstring for eval_tree_array(tree::Node{T}, cX::AbstractMatrix{T}, options::Options) where {T<:Real}. Check Documenter's build log for details.

Derivatives

SymbolicRegression.jl can automatically and efficiently compute derivatives of expressions with respect to variables or constants. This is done using either eval_diff_tree_array, to compute derivative with respect to a single variable, or with eval_grad_tree_array, to compute the gradient with respect all variables (or, all constants). Both use forward-mode automatic, but use Zygote.jl to compute derivatives of each operator, so this is very efficient.

Missing docstring.

Missing docstring for eval_diff_tree_array(tree::Node{T}, cX::AbstractMatrix{T}, options::Options, direction::Int) where {T<:Real}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eval_grad_tree_array(tree::Node{T}, cX::AbstractMatrix{T}, options::Options; variable::Bool=false) where {T<:Real}. Check Documenter's build log for details.

SymbolicUtils.jl interface

Missing docstring.

Missing docstring for node_to_symbolic(tree::Node, options::Options; varMap::Union{Array{String, 1}, Nothing}=nothing, index_functions::Bool=false). Check Documenter's build log for details.

Pareto frontier

SymbolicRegression.HallOfFameModule.calculate_pareto_frontier — Method

calculate_pareto_frontier(X::AbstractMatrix{T}, y::AbstractVector{T},
                        hallOfFame::HallOfFame{T}, options::Options;
                        weights=nothing, varMap=nothing) where {T<:Real}

Compute the dominating Pareto frontier for a given hallOfFame. This is the list of equations where each equation has a better loss than all simpler equations.

SymbolicRegression.HallOfFameModule.calculate_pareto_frontier — Method

calculate_pareto_frontier(dataset::Dataset{T}, hallOfFame::HallOfFame{T},
                        options::Options) where {T<:Real}